Common part of Anomaly Classification and Anomaly Segmentation moved to separate file.
